# internal/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: UalaTwitter API
  description: |
    Simplified microblogging backend. No authentication - pass the caller's UUID in
    the `X-User-ID` header on every request that requires identity.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /tweets:
    post:
      tags: [tweets]
      summary: Post a tweet
      description: |
        Text must be 1–280 Unicode runes. Media items are URL references only
        (no file upload). The tweet ID encodes author + creation time.
      parameters:
        - $ref: "#/components/parameters/XUserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 280
                  example: "Hello world!"
                media:
                  type: array
                  items:
                    $ref: "#/components/schemas/MediaItem"
      responses:
        "201":
          description: Tweet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tweet"
        "400":
          $ref: "#/components/responses/BadRequest"
  /users/{id}/tweets:
    get:
      tags: [tweets]
      summary: Get a user's tweets
      description: Returns tweets newest-first with cursor-based pagination.
      parameters:
        - $ref: "#/components/parameters/UserIDPath"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Before"
      responses:
        "200":
          description: List of tweets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tweet"

  # ── Follows ──────────────────────────────────────────────────────────────
...
components:
  parameters:
    XUserID:
      name: X-User-ID
      in: header
      required: true
      description: UUID of the acting user (no authentication — header is trusted).
      schema:
        type: string
        format: uuid
        example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    UserIDPath:
      name: id
      in: path
      required: true
      description: Target user UUID.
      schema:
        type: string
        format: uuid
        example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
    Limit:
      name: limit
      in: query
      description: Maximum number of results (1–50, default 20).
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 50
    Before:
      name: before
      in: query
      description: |
        Cursor for pagination. Pass the `id` of the last tweet from the previous
        page to retrieve the next page.
      schema:
        type: string
        example: "a1b2c3d4e5f67890abcdef1234567890_1713434400000"
  schemas:
...
    Tweet:
      type: object
      properties:
        id:
          type: string
          description: |
            Composite ID encoding author UUID and creation timestamp:
            `{32-char-hex-uuid}_{unix-milliseconds}`
          example: "a1b2c3d4e5f67890abcdef1234567890_1713434400000"
        user_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        text:
          type: string
          maxLength: 280
          example: "Hello world!"
        media:
          type: array
          items:
            $ref: "#/components/schemas/MediaItem"
        created_at:
          type: string
          format: date-time
    MediaItem:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [image, video, music]
          example: image
        url:
          type: string
          format: uri
          example: "https://example.com/photo.jpg"
  responses:
    BadRequest:
      description: Invalid input (missing field, empty tweet text, text over 280 chars, missing X-User-ID header).
      content:
        text/plain:
          schema:
            type: string
          examples:
            missingHeader:
              value: "missing X-User-ID header"
            emptyText:
              value: "tweet text cannot be empty"
            textTooLong:
              value: "text exceeds 280 characters"
            usernameRequired:
              value: "username required"
...